QBS Change card history histogram y-axis to be 100, 50, 25 or 15 max

+++ QBS/include/QBS_charth.h
@@ -36,7 +36,7 @@ protected:
 
     void draw_y_axis();
 
-    double max_;
+    int max_;
 
 };
